% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_buffers.R
\name{add_buffers}
\alias{add_buffers}
\title{Adding buffers to point-only data}
\usage{
add_buffers(
  data,
  crs = 4269,
  buffer_radius = 500,
  view = FALSE,
  file_name = NULL,
  output_dir = "outdata"
)
}
\arguments{
\item{data}{a tabular data that contains covariates and spatial informationf}

\item{crs}{Coordinate Reference System used for projection}

\item{buffer_radius}{radius in meters in which to create buffers}

\item{view}{a logical argument to specify whether the resulting data should be visualized}

\item{file_name}{the filename in which the exported object will be; if NULL then nothing is exported}

\item{output_dir}{directory name in which to export the file; "outdata" is set as the default}
}
\value{
if view == TRUE, then a map of output is returned. Otherwise an sf object containing the ID and geometry of points are returned.
}
\description{
The function takes in tabular data that contains geographic coordinates of point samples, projects it onto a coordinates system (default is ESPG:4269), and creates a buffer around each point. If view = TRUE, then a the buffers are visualized on a map. If a filename is provided, then the processes file is saved into the output_dir directory.
}
