\name{ lookup }

\alias{country}
\alias{adm}
\alias{alt}


\title{ lookup }

\description{ .. .. 
}

\usage{
country(lonlat, radius=0)
adm(lonlat, radius=0, maxrows=1)
alt(lonlat)
}


\arguments{
  \item{lonlat}{ }
  \item{radius}{  }
  \item{maxrows}{  }
}

\value{
}

\author{ Robert J. Hijmans}

\keyword{ spatial }

