\name{georefs-package}
\alias{georefs-package}
\alias{georefs}

\docType{package}

\title{
What the package does (short line)
}

\description{
More about what it does (maybe more than one line)
}

\details{
\tabular{ll}{
Package: \tab georefs\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-12-18\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Robert J. Hijmans and Jacob van Etten
}


\keyword{ package }
\keyword{ spatial }

