% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_coordinates.R
\name{multiple_coordinates}
\alias{multiple_coordinates}
\title{Constructing convex hulls}
\usage{
multiple_coordinates(data, group, crs = 4269)
}
\arguments{
\item{data}{a tabular data that contains Lat/Long coordinate information}

\item{group}{the grouping variable in data that should be used to compute convex hulls}

\item{crs}{the coordinate system that data should be projected onto}
}
\value{
a grouped tibble with grouping variable, nested version of original data, and respective convex hull geometry information
}
\description{
Studies that uses multiple coordinates to specifies its spatial extent can be aggregated by combining these coordinates into a single polygon described by the convex hull of these points. This function takes in data containing coordinate information and a grouping variable and computes a convex hull for each group.
}
\examples{
convex_hull(data = data, group = 'group')
}
