% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_setup.R
\name{osf_setup}
\alias{osf_setup}
\title{OSF setup}
\usage{
osf_setup(project_ID, PAT = NULL)
}
\arguments{
\item{project_ID}{a character string specifying the OSF identifier of the project}

\item{PAT}{OSF personal access token}
}
\description{
This function sets up a connection between your current R session and your project repository on OSF. If a Personal Access Token (PAT) is not provided in the argument, then a prompt will appear and ask for it.
}
\examples{
project <- osf_setup(project_ID = "osf_setup")
}
