% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_updating.R
\name{osf_updating}
\alias{osf_updating}
\title{OSF updating}
\usage{
osf_updating(
  project,
  component,
  type = c("download", "upload"),
  files = c("outdata", "rawdata", "tabular_shapefiles"),
  overwrite = TRUE
)
}
\arguments{
\item{project}{an osf_tbl object of the project}

\item{component}{the component in which files should be accessed from}

\item{type}{if "download" then all files on osf are downloaded ,if "upload",}

\item{files}{name of files to be uploaded to OSF; these could names of the file in the working directory or paths}

\item{overwrite}{logical argument specifying whether files should be overwritten}
}
\description{
This function allows you to download and upload files directly from and to your remote OSF repository
}
\examples{
project <- osf_setup(project_ID = "osf_setup")
osf_updating(project = project, component = "Data", type = 'download')
}
