% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_stats.R
\name{spatial_statistics}
\alias{spatial_statistics}
\title{Calculating spatial statistics}
\usage{
spatial_statistics(sf_object, view = FALSE)
}
\arguments{
\item{sf_object}{an sf object containing information of your shapefile}

\item{map}{a logical arugment specifying whether the processed data should be visualized on a map}
}
\value{
if view == TRUE, then a map of data is returned. Otherwise a table of calculated statistics is returned
}
\description{
The function takes in spatial information and calculates its area, centroid positions, and area bin class.
}
\examples{
spatial_statistics(sf_object)
}
