% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_polygons.R
\name{subset_polygons}
\alias{subset_polygons}
\title{Subsetting polygons from shape files}
\usage{
subset_polygons(
  sf_object,
  type = c("Parks", "Watershed", "Cities"),
  polygon_name = NULL,
  view = FALSE
)
}
\arguments{
\item{sf_object}{an sf object containing information of your shapefile}

\item{type}{the type of spatial information of your sf object; currently includes Parks, Watershed, and Cities}

\item{polygon_name}{the names of polygons in which you want to subset; if none included then the whole sf object is returned}

\item{view}{a logical arugment in which to specify whether the data should be visualized on a map}
}
\value{
if view == TRUE, then a map of the polygons is returned. Otherwise an sf object is returned containing a subset of the original polygons
}
\description{
This function allows you to subset polygons from other data sources (e.g. National Parks, Watersheds, and Cities).
}
\examples{
census <- here::here("tabular_shapefiles/KML_SHP/CensusSubdivisions_lcsd000b16a_e.shp")\%>\%
sf::st_read()
polygons <- census$CSDNAME\%>\%
sample(10)
subset_polygons(census, type = 'Cities', polygon_name = polygons, view = TRUE)
}
